requirements.txt
anthropic>=0.25.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pymongo>=4.6.0
dnspython>=2.4.0
